requirements.txt
absl-py==1.4.0
accelerate==0.21.0
addict==2.4.0
aiofiles==23.1.0
aiohttp==3.8.5
aiosignal==1.3.1
albumentations==0.4.3
altair==5.0.1
annotated-types==0.5.0
antlr4-python3-runtime==4.8
anyio==3.7.1
appdirs==1.4.4
async-timeout==4.0.2
attrs==23.1.0
basicsr==1.4.2
blinker==1.6.2
boltons==23.0.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
clean-fid==0.1.35
click==8.1.6
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
clip-anytorch==2.5.2
contourpy==1.1.0
cycler==0.11.0
decorator==5.1.1
diffusers==0.19.3
docker-pycreds==0.4.0
einops==0.3.0
facexlib==0.3.0
fastapi==0.101.0
ffmpy==0.3.1
filelock==3.12.2
filterpy==1.4.5
fonttools==4.42.0
frozenlist==1.4.0
fsspec==2023.6.0
ftfy==6.1.1
future==0.18.3
-e git+https://github.com/TencentARC/GFPGAN@2eac2033893ca7f427f4035d80fe95b92649ac56#egg=gfpgan
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.22.0
google-auth-oauthlib==1.0.0
gradio==3.39.0
gradio_client==0.3.0
grpcio==1.56.2
gunicorn==21.2.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.2
imgaug==0.2.6
importlib-metadata==6.8.0
iniconfig==2.0.0
invisible-watermark==0.2.0
Jinja2==3.1.2
jsonmerge==1.9.2
jsonschema==4.18.6
jsonschema-specifications==2023.7.1
k-diffusion==0.0.15
kiwisolver==1.4.4
kornia==0.6.0
lazy_loader==0.3
lightning-utilities==0.9.0
linkify-it-py==2.0.2
llvmlite==0.40.1
lmdb==1.4.1
Markdown==3.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mdit-py-plugins==0.3.3
mdurl==0.1.2
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numba==0.57.1
numpy==1.24.4
oauthlib==3.2.2
omegaconf==2.0.0
opencv-python==4.8.0.74
opencv-python-headless==4.8.0.74
orjson==3.9.4
packaging==23.1
pandas==2.0.3
pathtools==0.1.2
Pillow==9.5.0
platformdirs==3.10.0
pluggy==1.2.0
protobuf==4.23.4
psutil==5.9.5
pudb==2019.2
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.1.1
pydantic_core==2.4.0
pydeck==0.8.0
pyDeprecate==0.3.1
pydub==0.25.1
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.0.9
pytest==7.4.0
python-dateutil==2.8.2
python-multipart==0.0.6
pytorch-lightning==1.4.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0.1
-e git+https://github.com/xinntao/Real-ESRGAN@5ca1078535923d485892caee7d7804380bfc87fd#egg=realesrgan
referencing==0.30.2
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
resize-right==0.0.2
rich==13.5.2
rpds-py==0.9.2
rsa==4.9
safetensors==0.3.1
scikit-image==0.21.0
scipy==1.11.1
semantic-version==2.10.0
sentry-sdk==1.29.2
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
starlette==0.27.0
streamlit==1.25.0
sympy==1.12
taming-transformers-rom1504==0.0.6
tb-nightly==2.14.0a20230807
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
test-tube==0.7.5
tifffile==2023.7.18
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
torch-fidelity==0.3.0
torchdiffeq==0.2.3
torchmetrics==0.6.0
torchsde==0.2.5
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
trampoline==0.1.2
transformers==4.31.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
uc-micro-py==1.0.2
urllib3==1.26.16
urwid==2.1.2
uvicorn==0.23.2
validators==0.20.0
wandb==0.15.8
wcwidth==0.2.6
websockets==11.0.3
Werkzeug==2.3.6
yapf==0.40.1
yarl==1.9.2
zipp==3.16.2
